Conflicts:
	analysis/WhitespaceTokenizer.cpp
	contrib/analyzers/common/analysis/br/BrazilianStemmer.cpp
	contrib/analyzers/common/analysis/cjk/CJKTokenizer.cpp
	include/Lucene.h
	store/IndexInput.cpp
	test/index/IndexWriterTest.cpp
	util/Base64.cpp
	util/CharFolder.cpp
	util/StringUtils.cpp
	util/UnicodeUtils.cpp

+++ include/Collection.h
@@ -7,6 +7,9 @@
 #pragma once
 
 #include "LuceneObject.h"
+#include <set>
+#include <map>
+#include <vector>
 
 namespace Lucene
 {
@@ -204,7 +207,7 @@ namespace Lucene
 		
 		virtual int32_t hashCode()
 		{
-			return (int32_t)container.get();
+			return (int64_t)container.get();
 		}
 		
 		void swap(this_type& other)
